+++ libsolidity/analysis/DocStringTagParser.cpp
@@ -28,6 +28,8 @@
 #include <libsolidity/analysis/NameAndTypeResolver.h>
 #include <liblangutil/ErrorReporter.h>
 
+#include <boost/algorithm/string.hpp>
+
 using namespace std;
 using namespace solidity;
 using namespace solidity::langutil;
@@ -153,15 +155,17 @@ void DocStringTagParser::parseDocStrings(
 	_annotation.docTags = DocStringParser{*_node.documentation(), m_errorReporter}.parse();
 
 	size_t returnTagsVisited = 0;
-	for (auto const& docTag: _annotation.docTags)
+	for (auto const& [tagName, tagValue]: _annotation.docTags)
 	{
-		if (!_validTags.count(docTag.first))
+		if (boost::starts_with(tagName, "custom:") && tagName.size() > string("custom:").size())
+			continue;
+		else if (!_validTags.count(tagName))
 			m_errorReporter.docstringParsingError(
 				6546_error,
 				_node.documentation()->location(),
-				"Documentation tag @" + docTag.first + " not valid for " + _nodeName + "."
+				"Documentation tag @" + tagName + " not valid for " + _nodeName + "."
 			);
-		else if (docTag.first == "return")
+		else if (tagName == "return")
 		{
 			returnTagsVisited++;
 			if (auto const* varDecl = dynamic_cast<VariableDeclaration const*>(&_node))
@@ -171,19 +175,19 @@ void DocStringTagParser::parseDocStrings(
 					m_errorReporter.docstringParsingError(
 						5256_error,
 						_node.documentation()->location(),
-						"Documentation tag \"@" + docTag.first + "\" is only allowed once on state-variables."
+						"Documentation tag \"@" + tagName + "\" is only allowed once on state-variables."
 					);
 			}
 			else if (auto const* function = dynamic_cast<FunctionDefinition const*>(&_node))
 			{
-				string content = docTag.second.content;
+				string content = tagValue.content;
 				string firstWord = content.substr(0, content.find_first_of(" \t"));
 
 				if (returnTagsVisited > function->returnParameters().size())
 					m_errorReporter.docstringParsingError(
 						2604_error,
 						_node.documentation()->location(),
-						"Documentation tag \"@" + docTag.first + " " + docTag.second.content + "\"" +
+						"Documentation tag \"@" + tagName + " " + tagValue.content + "\"" +
 						" exceeds the number of return parameters."
 					);
 				else
@@ -193,7 +197,7 @@ void DocStringTagParser::parseDocStrings(
 						m_errorReporter.docstringParsingError(
 							5856_error,
 							_node.documentation()->location(),
-							"Documentation tag \"@" + docTag.first + " " + docTag.second.content + "\"" +
+							"Documentation tag \"@" + tagName + " " + tagValue.content + "\"" +
 							" does not contain the name of its return parameter."
 						);
 				}